Revert "[Explicit Module Builds] Serialize the inter-module dependency graph as incremental build state"
This reverts commit 9c6e1371944dc797018726acd4cd3167a078c1d8.
