+++ Sources/SwiftDriver/Driver/Driver.swift
@@ -235,6 +235,22 @@ public struct Driver {
   /// Only present when the driver will be writing the record.
   /// Only used for reading when compiling incrementally.
   @_spi(Testing) public let buildRecordInfo: BuildRecordInfo?
+
+  /// A build-record-relative path to the location of a serialized copy of the
+  /// driver's dependency graph.
+  ///
+  /// FIXME: This is a little ridiculous. We could probably just replace the
+  /// build record outright with a serialized format.
+  var driverDependencyGraphPath: VirtualPath? {
+    guard let recordInfo = self.buildRecordInfo else {
+      return nil
+    }
+    let filename = recordInfo.buildRecordPath.basenameWithoutExt
+    return recordInfo
+      .buildRecordPath
+      .parentDirectory
+      .appending(component: filename + ".priors")
+  }
   
   /// Whether to consider incremental compilation.
   let shouldAttemptIncrementalCompilation: Bool
@@ -1267,6 +1283,10 @@ extension Driver {
   public mutating func run(
     jobs: [Job]
   ) throws {
+    if parsedOptions.hasArgument(.v) {
+      try printVersion(outputStream: &stderrStream)
+    }
+
     let forceResponseFiles = parsedOptions.contains(.driverForceResponseFiles)
 
     // If we're only supposed to print the jobs, do so now.
@@ -1430,15 +1450,6 @@ extension Driver {
           /// Ensure that a bogus dependency graph is not used next time.
           buildRecordInfo?.removeBuildRecord()
           return
-      }
-      do {
-        try incrementalCompilationState.writeInterModuleDependencyGraph(buildRecordInfo)
-      }
-      catch {
-        diagnosticEngine.emit(
-          .warning("next compile must run a full dependency scan; could not write inter-module dependency graph: \(error.localizedDescription)"))
-        buildRecordInfo?.removeInterModuleDependencyGraph()
-        return
       }
     }
     buildRecordInfo?.writeBuildRecord(
