+++ Sources/SwiftDriver/Utilities/VirtualPath.swift
@@ -597,8 +597,6 @@ public struct TextualVirtualPath: Codable, Hashable {
       preconditionFailure("Path does not have a round-trippable textual representation")
     }
   }
-
-  internal var description: String { VirtualPath.lookup(path).description }
 }
 
 extension VirtualPath: CustomStringConvertible {
